requirements/pypi.txt
django-storages
boto3
pillow
djangorestframework
djangorestframework-filters
markdown
django-filter
